ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
xgboost==2.0.2
prophet==1.1.5
ortools==9.8.3296
python-dotenv==1.0.0
httpx==0.25.2